[CHEF-1804] add invalidation to sub-trees of merged attributes
Chef::Node::Attribute maintains a serial number which is incremented for
each reset. Merged attributes keep the serial number when they're
created, and check before reading.

This catches the case where a user has a reference to a now possibly
wrong part of the merged attributes, like this:

  node.default[:nested] = {:k => "old"}
  nested = node[:nested]
  node.default[:nested][:k] = "new"

`nested[:k]` would return "old" if cache invalidation was not
propagated.